requirements.txt
paddlepaddle==2.6.0
paddleocr==2.7.3
discord.py==2.3.2
python-dotenv==1.0.0
pillow==10.2.0
numpy<2.0.0
requests==2.31.0